extract static excluded lists and related imports to separate js file
